pyproject.toml
[project]
name = "coolifyeasyapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "google-cloud-storage>=2.10.0",
    "google-auth>=2.23.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "cryptography>=41.0.0",
    "requests>=2.31.0",
    "google-cloud-speech>=2.34.0",
]
